1. corrected VPN service status on GUI
